+++ include/fdf.h
@@ -6,7 +6,7 @@
 /*   By: ksoto <ksoto@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/10 18:33:16 by ciglesia          #+#    #+#             */
-/*   Updated: 2021/10/03 22:12:14 by ksoto            ###   ########.fr       */
+/*   Updated: 2021/10/13 03:50:54 by ksoto            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,8 +54,11 @@ typedef struct s_fdf
 	int				res[2];
 	int				view;
 	int				color;
+	int				color1;
 	int				shift_x;
 	int				shift_y;
+	float			diff_x;
+	float			diff_y;
 	int				zoom;
 	int				**bmp;
 	t_pixel			*init;
@@ -68,6 +71,8 @@ typedef struct s_fdf
 
 t_fdf	*ft_fdf(char *name);
 int		exit_win(t_fdf	*fdf);
+void	init_map(t_fdf	*fdf);
+int		key_win(int key, t_fdf *fdf);
 
 /*
 **	Display
@@ -88,10 +93,11 @@ void	free_map(void);
 /*
 ** Compute
 */
+void	bresenham_line(t_fdf *fdf, int **bmp);
 void	set_coord(t_pixel *pixel, float x, float y);
 void	set_horizontal(t_fdf *fdf, int x, int y);
 void	set_vertical(t_fdf *fdf, int x, int y);
-void	set_color(t_fdf *fdf);
+void	set_color(t_fdf *fdf, int **bmp, int z, int z1);
 void	convert_isometric(t_pixel *px, int z, t_fdf *fdf);
 int		module(int a);
 int		max_calculator(int a, int b);
