+++ lib/libft/src/ft_atoi_base.c
@@ -1,0 +1,51 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_atoi_base.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ksoto <ksoto@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/10/13 03:58:14 by ksoto             #+#    #+#             */
+/*   Updated: 2021/10/13 03:59:38 by ksoto            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libft.h"
+
+int	nbr_inbase(char c, int base)
+{
+	if (base <= 10)
+		return (c >= '0' && c <= '9');
+	return ((c >= '0' && c <= '9') || (c >= 'A' && c <= ('A' + base - 10)) || \
+	(c >= 'a' && c <= ('a' + base - 10)));
+}
+
+int	ft_atoi_base(const char *str, int base)
+{
+	int		i;
+	int		nbr;
+	int		sign;
+
+	if (!str[0] || (base < 2 || base > 16))
+		return (0);
+	i = 0;
+	nbr = 0;
+	sign = 1;
+	while (str[i] == '\t' || str[i] == '\v' || str[i] == '\n' || \
+		str[i] == ' ' || str[i] == '\r' || str[i] == '\f')
+		i += 1;
+	if (str[i] == '-' || str[i] == '+')
+		if (str[i++] == '-')
+			sign *= -1;
+	while (str[i] && nbr_inbase(str[i], base))
+	{
+		if (str[i] >= 'A' && 'F' >= str[i])
+			nbr = (nbr * base) + (str[i] - 'A' + 10);
+		else if (str[i] >= 'a' && 'f' >= str[i])
+			nbr = (nbr * base) + (str[i] - 'a' + 10);
+		else
+			nbr = (nbr * base) + (str[i] - '0');
+		i += 1;
+	}
+	return (nbr * sign);
+}
