+++ src/display/render.c
@@ -6,7 +6,7 @@
 /*   By: ksoto <ksoto@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/11 01:27:49 by ciglesia          #+#    #+#             */
-/*   Updated: 2021/10/13 02:48:43 by ksoto            ###   ########.fr       */
+/*   Updated: 2021/10/13 04:22:15 by ksoto            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ static void	fill_img(t_layer *l, int w, int h, int **bmp)
 	int				y;
 	int				color;
 	unsigned char	*ptr;
-	t_fdf			*fdf;
+	t_fdf			*fdf;//cambiar
 
 	fdf = ft_fdf(NULL);
 	y = 0;
@@ -74,7 +74,8 @@ void	ft_plot(int **bmp)
 	fdf = ft_fdf(NULL);
 	if (fdf->l.img)
 		mlx_destroy_image(fdf->mlx, fdf->l.img);
-	if (!(l.img = mlx_new_image(fdf->mlx, fdf->res[0], fdf->res[1])))
+	l.img = mlx_new_image(fdf->mlx, fdf->res[0], fdf->res[1]);
+	if (!l.img)
 		exit_win(ft_fdf(NULL));
 	l.data = mlx_get_data_addr(l.img, &l.bpp, &l.bpl, &l.endian);
 	fill_img(&l, fdf->res[0], fdf->res[1], bmp);
@@ -83,52 +84,32 @@ void	ft_plot(int **bmp)
 }
 
 /*
-** bresenham_line: bresenham algorithm to plot a line pixel by pixel
-** diff_x: distance that x need to advance
-** diff_y: distance that y need to advance
+** bresenham_loop: bresenham loop
+** fdf: fdf project
 */
 
-void	bresenham_line(t_fdf *fdf, int **bmp)
+void	bresenham_loop(t_fdf *fdf)
 {
-	float	diff_x;
-	float	diff_y;
-	int		max;
-	long	z;
-	long	z1;
-	int		color;
-	int		color1;
+	int	x;
+	int	y;
 
-	z = fdf->map[(int)fdf->init->y][(int)fdf->init->x].z;
-	z1 = fdf->map[(int)fdf->end->y][(int)fdf->end->x].z;
-	color = fdf->map[(int)fdf->init->y][(int)fdf->init->x].color;
-	color1 = fdf->map[(int)fdf->init->y][(int)fdf->init->x].color;
-	convert_zoom(fdf);
-	convert_isometric(fdf->init, z, fdf);
-	convert_isometric(fdf->end, z1, fdf);
-	convert_shift(fdf);
-	diff_x = fdf->end->x - fdf->init->x;
-	diff_y = fdf->end->y - fdf->init->y;
-	max = max_calculator(module(diff_x), module(diff_y));
-	diff_x /= max;
-	diff_y /= max;
-	while (((int)(fdf->init->x - fdf->end->x) || \
-			(int)(fdf->init->y - fdf->end->y)) && \
-			((int)fdf->init->y > 0) && \
-			((int)fdf->init->x > 0) && \
-			((int)fdf->init->y < fdf->res[1]) && \
-			((int)fdf->init->x < fdf->res[0])
-			)
+	y = -1;
+	while (++y < fdf->mapy)
 	{
-		if (z || z1)
-			bmp[(int)fdf->init->y][(int)fdf->init->x] = 0xFFFFFF;
-		else
-			bmp[(int)fdf->init->y][(int)fdf->init->x] = 0x00FF00;
-		if (color)
-			bmp[(int)fdf->init->y][(int)fdf->init->x] = color;
-		else if (color1)
-			bmp[(int)fdf->init->y][(int)fdf->init->x] = color1;
-		fdf->init->x += diff_x;
-		fdf->init->y += diff_y;
+		x = -1;
+		while (++x < fdf->mapx)
+		{
+			if (x < fdf->mapx - 1)
+			{
+				set_vertical(fdf, x, y);
+				bresenham_line(fdf, fdf->bmp);
+			}
+			if (y < fdf->mapy - 1)
+			{
+				set_horizontal(fdf, x, y);
+				bresenham_line(fdf, fdf->bmp);
+			}
+		}
 	}
 }
 
@@ -150,23 +131,6 @@ void	plot_map(t_fdf *fdf)
 			fdf->bmp[y][x] = 0;
 	}
 	mlx_clear_window(fdf->mlx, fdf->win);
-	y = -1;
-	while (++y < fdf->mapy)
-	{
-		x = -1;
-		while (++x < fdf->mapx)
-		{
-			if (x < fdf->mapx - 1)
-			{
-				set_vertical(fdf, x, y);
-				bresenham_line(fdf, fdf->bmp);
-			}
-			if (y < fdf->mapy - 1)
-			{
-				set_horizontal(fdf, x, y);
-				bresenham_line(fdf, fdf->bmp);
-			}
-		}
-	}
+	bresenham_loop(fdf);
 	ft_plot(fdf->bmp);
 }
